added optimized cuda version added pstl version with native ordering added (renamed) optimized pstl version
